requirements.txt
fastapi==0.115.6
uvicorn==0.32.0
psycopg[binary]==3.2.3
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
